requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
cycler==0.10.0
decorator==4.4.2
gast==0.4.0
grpcio==1.35.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==3.4.0
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
mock==4.0.3
networkx==2.5
numpy==1.19.5
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tftb==0.1.1
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.56.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
